+++ minimap.c
@@ -6,7 +6,7 @@
 /*   By: thisai <thisai@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/21 18:20:58 by thisai            #+#    #+#             */
-/*   Updated: 2021/02/24 16:55:16 by thisai           ###   ########.fr       */
+/*   Updated: 2021/02/24 18:20:38 by thisai           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -95,47 +95,4 @@ void	c3_draw_map(t_c3_state *stat)
 		i++;
 	}
 	c3_draw_player_on_map(stat);
-}
-
-void	c3_draw_rays_on_map(t_c3_state *stat)
-{
-	double		world_x;
-	double		world_y;
-	double		screen_x;
-	double		screen_y;
-	int			x;
-	int			i;
-
-	x = 0;
-	while (x < stat->renderer.resolution_x)
-	{
-		world_x = stat->renderer.rays[x].hits[0].position.x;
-		world_y = stat->renderer.rays[x].hits[0].position.y;
-		screen_x = world_x * stat->renderer.minimap_width / stat->map.width;
-		screen_y = world_y * stat->renderer.minimap_height / stat->map.height;
-
-		int r = 255 * x / stat->renderer.resolution_x;
-		int col = (r << 16) + ((255 - r) << 0);
-		mlx_string_put(
-			stat->mlx, stat->window, screen_x, screen_y,
-			mlx_get_color_value(stat->mlx, col), "*");
-
-		i = 0;
-		while (i < stat->renderer.rays[x].hit_sprite_count)
-		{
-			world_x = stat->renderer.rays[x].hits[i + 1].position.x;
-			world_y = stat->renderer.rays[x].hits[i + 1].position.y;
-			screen_x = world_x * stat->renderer.minimap_width / stat->map.width;
-			screen_y = world_y * stat->renderer.minimap_height / stat->map.height;
-
-			int r = 255 * x / stat->renderer.resolution_x;
-			int col = (r << 16) + ((255 - r) << 0);
-			mlx_string_put(
-				stat->mlx, stat->window, screen_x, screen_y,
-				mlx_get_color_value(stat->mlx, col), "x");
-			i++;
-		}
-
-		x++;
-	}
 }
